requirements.txt
fastapi
uvicorn
python-docx
docx2pdf
num2words
pydantic
